SMLMData

Documentation for SMLMData.

SMLMData.SMLD2D

SMLMData.SMLD2D

Base.length

SMLMData.addoffset

SMLMData.addoffset!

SMLMData.catsmld

SMLMData.isolateconnected

SMLMData.isolatesmld

SMLMData.isolatesmld

SMLMData.isolatesmld

SMLMData.SMLD2D — Method

SMLD2D(nlocalizations::Int)

Constructor to generate an smld from a data frame.

Description

This is a constructor for the SMLD2D struct which allows you to populate the structure with data defined in the dataframe data. The intention is that a .csv table can be organized with localizations on each row as [datasetnum, framenum, x, y, σx, σy], loaded using the CSV package, and placed into a dataframe with the DataFrames package.

SMLMData.SMLD2D — Method

SMLD2D(nlocalizations::Int)

Constructor to generate an empty smld with a specific size.

Description

This is a constructor for the SMLD2D struct which allows you to populate the structure with undefined values for a predefined number of localizations.

Base.length — Method

length(smld::SMLMData.SMLD2D)

Returns the number of localizations in smld.

SMLMData.addoffset! — Method

addoffset!(smld::SMLD2D, offset::Vector{Float64})

Add an offset elementwise to each localization smld.y and smld.x.

Inputs

-offset: Two element vector describing the offset. ([offsety; offsetx])

SMLMData.addoffset — Method

smld_offset = addoffset(smld::SMLD2D, offset::Vector{Float64})

Add an offset elementwise to each localization smld.y and smld.x.

Inputs

-offset: Two element vector describing the offset. ([offsety; offsetx])

SMLMData.catsmld — Method

smld = catsmld(smld1::SMLD2D, smld2::SMLD2D)

Concatenate smld1 and smld2.

Description

This method concatenates the datafields of smld1 and smld2 (as defined by smld1.datafields) into an output smld. The non-data fields are taken from smld1 and copied into smld.

SMLMData.isolateconnected — Method

smld_connected, keepbool = isolateconnected(smld::SMLMData.SMLD2D)

Prepare individual SMLD2D structures for each precluster in smld.

Description

Connected localizations in smld (based on their connectID field) are reorganized into distinct SMLD2D structures, with each structure corresponding to a single connectID.

Inputs

-smld: SMLD2D structure containing the localization data, with the field connectID populated with a meaningful set of connection indices.

outputs

-smld_connected: Vector of SMLD2D structures with each structure containing localizations that share the same connectID. -keepbool: Vector of BitVector defining the smld indices sharing the same connectID.

SMLMData.isolatesmld — Method

smld_sub = isolatesmld(smld::SMLD2D, keepbit::BitVector)

Isolate the smld localizations specified by keepbit BitVector.

Description

This method grabs the localizations requested by the BitVector keepbit, which should have the same length as the number of localizations in smld.

SMLMData.isolatesmld — Method

smld_sub = isolatesmld(smld::SMLD2D, subind::UnitRange)

Isolate the smld localizations specified by subind indices.

Description

This method grabs the localizations indexed by subind from smld and outputs a new SMLD2D structure containing only those localization.

SMLMData.isolatesmld — Method

smld_sub = isolatesmld(smld::SMLD2D, subind::Vector{Int})

Isolate the smld localizations specified by subind indices.

Description

This method grabs the localizations indexed by subind from smld and outputs a new SMLD2D structure containing only those localization.