SMLMData

Documentation for SMLMData.

SMLMData.SMLD2D

SMLMData.SMLD2D

Base.length

SMLMData.addoffset

SMLMData.addoffset!

SMLMData.catsmld

SMLMData.isolateconnected

SMLMData.isolatesmld

SMLMData.isolatesmld

SMLMData.isolatesmld

SMLMData.makebinim

SMLMData.makebinim

SMLMData.makecircleim

SMLMData.makecircleim

SMLMData.makegaussim

SMLMData.makegaussim

SMLMData.makehistim

SMLMData.makehistim

SMLMData.SMLD2D — Method

SMLD2D(nlocalizations::Int)

Constructor to generate an smld from a data frame.

Description

This is a constructor for the SMLD2D struct which allows you to populate the structure with data defined in the dataframe data. The intention is that a .csv table can be organized with localizations on each row as [datasetnum, framenum, x, y, σx, σy], loaded using the CSV package, and placed into a dataframe with the DataFrames package.

SMLMData.SMLD2D — Method

SMLD2D(nlocalizations::Int)

Constructor to generate an empty smld with a specific size.

Description

This is a constructor for the SMLD2D struct which allows you to populate the structure with undefined values for a predefined number of localizations.

Base.length — Method

length(smld::SMLMData.SMLD2D)

Returns the number of localizations in smld.

SMLMData.addoffset! — Method

addoffset!(smld::SMLD2D, offset::Vector{Float64})

Add an offset elementwise to each localization smld.y and smld.x.

Inputs

-offset: Two element vector describing the offset. ([offsety; offsetx])

SMLMData.addoffset — Method

smld_offset = addoffset(smld::SMLD2D, offset::Vector{Float64})

Add an offset elementwise to each localization smld.y and smld.x.

Inputs

-offset: Two element vector describing the offset. ([offsety; offsetx])

SMLMData.catsmld — Method

smld = catsmld(smld1::SMLD2D, smld2::SMLD2D)

Concatenate smld1 and smld2.

Description

This method concatenates the datafields of smld1 and smld2 (as defined by smld1.datafields) into an output smld. The non-data fields are taken from smld1 and copied into smld.

SMLMData.isolateconnected — Method

smld_connected, keepbool = isolateconnected(smld::SMLMData.SMLD2D)

Prepare individual SMLD2D structures for each precluster in smld.

Description

Connected localizations in smld (based on their connectID field) are reorganized into distinct SMLD2D structures, with each structure corresponding to a single connectID.

Inputs

-smld: SMLD2D structure containing the localization data, with the field connectID populated with a meaningful set of connection indices.

outputs

-smld_connected: Vector of SMLD2D structures with each structure containing localizations that share the same connectID. -keepbool: Vector of BitVector defining the smld indices sharing the same connectID.

SMLMData.isolatesmld — Method

smld_sub = isolatesmld(smld::SMLD2D, keepbit::BitVector)

Isolate the smld localizations specified by keepbit BitVector.

Description

This method grabs the localizations requested by the BitVector keepbit, which should have the same length as the number of localizations in smld.

SMLMData.isolatesmld — Method

smld_sub = isolatesmld(smld::SMLD2D, subind::UnitRange)

Isolate the smld localizations specified by subind indices.

Description

This method grabs the localizations indexed by subind from smld and outputs a new SMLD2D structure containing only those localization.

SMLMData.isolatesmld — Method

smld_sub = isolatesmld(smld::SMLD2D, subind::Vector{Int})

Isolate the smld localizations specified by subind indices.

Description

This method grabs the localizations indexed by subind from smld and outputs a new SMLD2D structure containing only those localization.

SMLMData.makebinim — Function

image = makebinim(coords::Matrix{Float64}, 
                  datasize::Vector{Float64},
                  mag::Float64 = 20.0)

Make a binary image of the localizations in coords.

Description

This function creates an image of the localizations in coords by placing a hot pixel at the coordinates of each localization.

Inputs

-coords: Localization coordinates. ([y x]) -datasize: Size of the data image. ([ysize xsize]) -mag: Approximate magnfication from data coordinates to SR coordinates. (Default = 20.0)

Outputs

-image: Matrix{Float64} binary image of localizations.

SMLMData.makebinim — Function

image = makebinim(smld::SMLMData.SMLD2D, mag::Float64 = 20.0)

Make a binary image of the localizations in smld.

Description

This function creates an image of the localizations in smld by placing a hot pixel at the coordinates of each localization.

Inputs

-smld: SMLMData.SMLD2D data structure containing localizations. -mag: Approximate magnfication from data coordinates to SR coordinates. (Default = 20.0)

Outputs

-image: Matrix{Float64} binary image of localizations.

SMLMData.makecircleim — Function

image = makecircleim(smld::SMLMData.SMLD2D, mag::Float64 = 20.0)

Make a circle image of the localizations in smld.

Description

This function creates an image of the localizations in smld by adding a circle for each localization.

Inputs

-smld: SMLMData.SMLD2D data structure containing localizations. -mag: Approximate magnfication from data coordinates to SR coordinates. (Default = 20.0)

Outputs

-image: Matrix{Float64} circle image of localizations.

SMLMData.makecircleim — Function

image = makecircleim(coords::Matrix{Float64},
                     σ::Vector{Float64},
                     datasize::Vector{Int},
                     mag::Float64 = 20.0)

Make a circle image of the localizations in coords.

Description

This function creates an image of the localizations in coords by adding a circle centered at the locations coords with radii σ.

Inputs

-coords: Localization coordinates. ([y x]) -σ: Standard error of localizations in coords. (nlocx1) -datasize: Size of the data image. ([ysize xsize]) -mag: Approximate magnfication from data coordinates to SR coordinates. (Default = 20.0)

Outputs

-image: Matrix{Float64} histogram image of localizations.

SMLMData.makegaussim — Function

image = makegaussim(μ::Matrix{Float64},
                    σ_μ::Matrix{Float64}, 
                    photons::Vector{Float64},
                    datasize::Vector{Int},
                    mag::Float64 = 20.0, 
                    nsigma::Float64 = 5.0)

Make a Gaussian image of the localizations in μ.

Description

This function creates an image of the localizations defined by μ and σ_μ in which Gaussians with standard deviations σ_μ are added at positions μ.

Inputs

-μ: Positions of the Gaussians. (pixels)([y x]) -σ_μ: Standard errors of μ estimates. (pixels)([y x]) -photons: Photons attributed to each row of μ. -datasize: Size of the region of data collection. (pixels)([y; x]) -mag: Approximate magnfication from data coordinates to SR coordinates. (Default = 20.0) -nsigma: Number of standard deviations from the localization coordinate at which we truncate the Gaussian. (Default = 5.0)

Outputs

-image: Matrix{Float64} Gaussian image in which each localization in smld is plotted as a Gaussian.

SMLMData.makegaussim — Function

image = makegaussim(smld::SMLMData.SMLD2D, 
                    mag::Float64 = 20.0, 
                    nsigma::Float64 = 5.0)

Make a Gaussian image of the localizations in smld.

Description

This function creates an image of the localizations in smld by placing a Gaussian truncated to nsigma at the localization coordinates, where the standard deviation is given by smld.σ_x and smld.σ_y. The image is then normalized such that it sums to 1.0. The background signal is not accounted for in this method.

Inputs

-smld: SMLMData.SMLD2D data structure containing localizations. -mag: Approximate magnfication from data coordinates to SR coordinates. (Default = 20.0) -nsigma: Number of standard deviations from the localization coordinate at which we truncate the Gaussian. (Default = 5.0)

Outputs

-image: Matrix{Float64} Gaussian image in which each localization in smld is plotted as a Gaussian.

SMLMData.makehistim — Function

image = makehistim(coords::Matrix{Float64},
                   datasize::Vector{Int},
                   mag::Float64 = 20.0)

Make a histogram image of the localizations in coords.

Description

This function creates an image of the localizations in coords by adding 1.0 to a pixel for each localization present within that pixel. The final image is then scaled so that it sums to 1.0.

Inputs

-coords: Localization coordinates. ([y x]) -datasize: Size of the data image. ([ysize xsize]) -mag: Approximate magnfication from data coordinates to SR coordinates. (Default = 20.0)

Outputs

-image: Matrix{Float64} histogram image of localizations.

SMLMData.makehistim — Function

image = makehistim(smld::SMLMData.SMLD2D, mag::Float64 = 20.0)

Make a histogram image of the localizations in smld.

Description

This function creates an image of the localizations in smld by adding 1.0 to a pixel for each localization present within that pixel. The final image is then scaled so that it sums to 1.0.

Inputs

-smld: SMLMData.SMLD2D data structure containing localizations. -mag: Approximate magnfication from data coordinates to SR coordinates. (Default = 20.0)

Outputs

-image: Matrix{Float64} histogram image of localizations.